Move transactions off of dashboard
