+++ routes/main.js
@@ -2,12 +2,9 @@ const express = require("express");
 const router = express.Router();
 const authController = require("../controllers/auth");
 const homeController = require("../controllers/home");
-//const transactionsController = require("../controllers/transactions");
 const { ensureAuth, ensureGuest } = require("../middleware/auth");
 
-//Main Routes - simplified for now
 router.get("/", homeController.getIndex);
-//router.get("/transactions", ensureAuth, transactionsController.getTransactions);
 router.get("/login", authController.getLogin);
 router.post("/login", authController.postLogin);
 router.get("/logout", authController.logout);
