+++ views/partials/header.ejs
@@ -53,7 +53,9 @@
                 <div class="flex space-x-4">
                   <!-- Current: "bg-gray-900 text-white", Default: "text-gray-300 hover:bg-gray-700 hover:text-white" -->
                   <!-- <a href="/reports/dashboard" class="bg-primary-focus text-white px-3 py-2 rounded-md text-sm font-medium" aria-current="page">Dashboard</a> -->
-                  <a href="/reports/dashboard" class="text-base-300 hover:bg-primary-focus hover:text-base-100 px-3 py-2 rounded-md text-sm font-medium" aria-current="page">Dashboard</a>
+                  <a href="/dashboard" class="text-base-300 hover:bg-primary-focus hover:text-base-100 px-3 py-2 rounded-md text-sm font-medium" aria-current="page">Dashboard</a>
+
+                  <a href="/transactions" class="text-base-300 hover:bg-primary-focus hover:text-base-100 px-3 py-2 rounded-md text-sm font-medium" aria-current="page">Transactions</a>
       
                   <a href="/entities" class="text-base-300 hover:bg-primary-focus hover:text-base-100 px-3 py-2 rounded-md text-sm font-medium">Entities</a>
       
@@ -113,8 +115,10 @@
         <div id="mobile-menu" class="hidden" id="mobile-menu">
           <div class="space-y-1 px-2 pt-2 pb-3">
             <!-- Current: "bg-gray-900 text-white", Default: "text-gray-300 hover:bg-gray-700 hover:text-white" -->
-            <a href="/reports/dashboard" class="text-base-300 hover:bg-primary-focus hover:text-base-100 block px-3 py-2 rounded-md text-base font-medium" aria-current="page">Dashboard</a>
+            <a href="/dashboard" class="text-base-300 hover:bg-primary-focus hover:text-base-100 block px-3 py-2 rounded-md text-base font-medium" aria-current="page">Dashboard</a>
       
+            <a href="/transactions" class="text-base-300 hover:bg-primary-focus hover:text-base-100 block px-3 py-2 rounded-md text-base font-medium" aria-current="page">Transactions</a>
+
             <a href="/entities" class="text-base-300 hover:bg-primary-focus hover:text-base-100 block px-3 py-2 rounded-md text-base font-medium">Entities</a>
       
             <a href="/accounts" class="text-base-300 hover:bg-primary-focus hover:text-base-100 block px-3 py-2 rounded-md text-base font-medium">Accounts</a>
